% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unb_ativos.R
\name{unb_covariancia}
\alias{unb_covariancia}
\title{Covariância}
\usage{
unb_covariancia(prob, valor1, valor2)
}
\arguments{
\item{prob}{vetor de números.}

\item{valor1}{vetor de números.}

\item{valor2}{vetor de números.}
}
\description{
Calcula a covariância de dois vetores.
}
