% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unb_ativos.R
\name{unb_desvio_padrao}
\alias{unb_desvio_padrao}
\title{Desvio-padrão.}
\usage{
unb_desvio_padrao(prob, valor)
}
\arguments{
\item{prob}{vetor de números.}

\item{valor}{vetor de números.}
}
\description{
Calcula o desvio-padrão com base em um vetor numérico com respectivas probabilidades associadas.
}
