% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unb_ativos.R
\name{unb_esperanca}
\alias{unb_esperanca}
\title{Esperança.}
\usage{
unb_esperanca(prob, valor)
}
\arguments{
\item{prob}{vetor de números.}

\item{valor}{vetor de números.}
}
\description{
Calcula o valor esperado com base em um vetor numérico com respectivas probabilidades associadas.
}
