% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unb_var.R
\name{unb_incremental_var}
\alias{unb_incremental_var}
\title{Incremental VaR}
\usage{
unb_incremental_var(mapeamento, volatilidade, correlacao, alpha, incremento)
}
\arguments{
\item{mapeamento}{vetor de números.}

\item{volatilidade}{vetor de números.}

\item{correlacao}{vetor de números}

\item{alpha}{escalar entre 0 e 1.}
}
\description{
Recebe vetores de mapeamento, volatilidade, correlacao, nível de significância alpha e vetor de incremento e retorna escalar.
}
