% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unb_carteira.R
\name{unb_matriz_covar}
\alias{unb_matriz_covar}
\title{Matriz de variância-covariância dos ativos.}
\usage{
unb_matriz_covar(desviopadrao, correlacao)
}
\arguments{
\item{correlacao}{matriz de correlação.}

\item{diag_dp}{vetor de números.}
}
\description{
Recebe vetores de desvio padrão e matriz de correlação e calcula a matriz de variância-covariância.
}
