% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unb_sigma.R
\name{unb_q_corr}
\alias{unb_q_corr}
\title{Matriz Q}
\usage{
unb_q_corr(volatilidade, correlacao, alpha)
}
\arguments{
\item{volatilidade}{vetor de números.}

\item{correlacao}{vetor de números.}

\item{alpha}{escalar entre 0 e 1.}
}
\description{
Recebe vetores de volatilidade e correlação e nível de significância alpha e retorna matriz Q
}
