% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unb_sigma.R
\name{unb_q_cov}
\alias{unb_q_cov}
\title{Matriz Q}
\usage{
unb_q_cov(sigma, alpha)
}
\arguments{
\item{sigma}{matriz de variância-covariância.}

\item{alpha}{escalar entre 0 e 1.}
}
\description{
Recebe matriz de variância-covariância nível de significância alpha e retorna matriz Q
}
