% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unb_carteira.R
\name{unb_retorno_esp_c}
\alias{unb_retorno_esp_c}
\title{Retorno esperado da carteira.}
\usage{
unb_retorno_esp_c(peso, retorno_esp)
}
\arguments{
\item{peso}{vetor de números.}

\item{retorno_esp}{vetor de números.}
}
\description{
Recebe vetores de peso e retorno esperado e calcula o retorno esperado da carteira.
}
