% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unb_carteira.R
\name{unb_risco_tot_c}
\alias{unb_risco_tot_c}
\title{Risco total da carteira.}
\usage{
unb_risco_tot_c(peso, riscototal, correlacao)
}
\arguments{
\item{peso}{vetor de números.}

\item{riscototal}{vetor de números.}

\item{correlacao}{matriz de correlação.}
}
\description{
Recebe vetores de peso, risco total e matriz de correlação e calcula o risco total da carteira.
}
