% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unb_sigma.R
\name{unb_sigma}
\alias{unb_sigma}
\title{Matriz Sigma}
\usage{
unb_sigma(volatilidade, correlacao)
}
\arguments{
\item{volatilidade}{vetor de números.}

\item{correlacao}{vetor de números.}
}
\description{
Recebe vetores de volatilidade e correlação e retorna matriz de variância covariância
}
