% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unb_var.R
\name{unb_var}
\alias{unb_var}
\title{Value at Risk}
\usage{
unb_var(mapeamento, volatilidade, correlacao, alpha)
}
\arguments{
\item{mapeamento}{vetor de números.}

\item{volatilidade}{vetor de números.}

\item{correlacao}{vetor de números.}

\item{alpha}{escalar entre 0 e 1.}
}
\description{
Recebe vetores de mapeamento, volatilidade, correlacao e nível de significância alpha retorna value at risk.
}
