% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unb_var.R
\name{unb_var_u}
\alias{unb_var_u}
\title{Value at Risk não diversificado.}
\usage{
unb_var_u(mapeamento, volatilidade, alpha)
}
\arguments{
\item{mapeamento}{vetor de números.}

\item{volatilidade}{vetor de números.}

\item{alpha}{escalar entre 0 e 1.}
}
\description{
Recebe vetores de mapeamento, volatilidadee nível de significância alpha retorna value at risk.
}
