% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unb_ativos.R
\name{unb_variancia}
\alias{unb_variancia}
\title{Variância.}
\usage{
unb_variancia(prob, valor)
}
\arguments{
\item{prob}{vetor de números.}

\item{valor}{vetor de números.}
}
\description{
Calcula a variância com base em um vetor numérico com respectivas probabilidades associadas.
}
