% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unb_carteira.R
\name{unb_variancia_c}
\alias{unb_variancia_c}
\title{Variância do retorno da carteira.}
\usage{
unb_variancia_c(peso, riscototal, correlacao)
}
\arguments{
\item{peso}{vetor de números.}

\item{riscototal}{vetor de números.}

\item{correlacao}{matriz de correlação.}
}
\description{
Recebe vetores de peso, risco total e matriz de correlação e calcula a variância do retorno da carteira.
}
